fix import-then-export case
This fixes #33.

I opted to do a less-invasive change that only adjusts the behavior for the known-bad case of a local name that is being used as an ExportSpecifier.

Arguably, we should always use this new pattern, because it's not really safe to construct identifiers (or call `path.scope.rename`) with names like `Ember.String` (which is not a valid Javascript identifier, it needs to be represented as a MemberExpression).