Ensure identifiers are properly re-written.
The code previously took an identifier and replaced it's `name` value
with one that included a string like `Ember.Foo`. This is completely
invalid (a string like `Ember.Foo` is actually a MemberExpression), but
we got away with it because for the most part Babel happily printed
**whatever** value you give it in that spot.

Unfortunately, this happy mistake bit us fairly severely due to Babel
introducing new plugins that run as part of `@babel/preset-env` (e.g. to
make sure that all identifiers would work properly on IE9).

This commit changes from doing a `path.scope.rename(importName,
globalPathString)` to properly replacing the identifier with a member
expression.