Ensure decorators provided by Ember are handled when the identifier is aliased.
tldr; decorator paths are not always included in
`path.scope.getBinding(local.name)`

In some circumstances, decorators are not included in the reference
paths for a local binding when the decorator identifier name is also
defined _within_ the method being decorated. This is likely a bug in
Babel, that should be reported and fixed.