Make ensureImport lazy
Makes the logic for ensuring the Ember import lazy. This allows it to
handle cases where a different transform has also added the Ember import
already, and so it can use it lazily. Also makes the logic generalized
(in case we ever need to import from a different module in the future).